split up sender concepts and completion signature utilities into separate headers
