+++ include/stdexec/__detail/__basic_sender.hpp
@@ -21,6 +21,8 @@
 #include "__diagnostics.hpp"
 #include "__env.hpp"
 #include "__meta.hpp"
+#include "__senders.hpp"
+#include "__sender_introspection.hpp"
 #include "__tuple.hpp"
 #include "__type_traits.hpp"
 
@@ -43,32 +45,6 @@ namespace stdexec {
         return static_cast<_Data&&>(__data);
       }
     };
-
-    template <class _Tag, class _Data, class... _Child>
-    struct __desc {
-      using __tag = _Tag;
-      using __data = _Data;
-      using __children = __types<_Child...>;
-
-      template <class _Fn>
-      using __f = __minvoke<_Fn, _Tag, _Data, _Child...>;
-    };
-
-    template <class _Fn>
-    struct __sexpr_uncurry_fn {
-      template <class _Tag, class _Data, class... _Child>
-      constexpr auto operator()(_Tag, _Data&&, _Child&&...) const noexcept
-        -> __minvoke<_Fn, _Tag, _Data, _Child...>;
-    };
-
-    template <class _Sender, class _Fn>
-    using __sexpr_uncurry =
-      __call_result_t<__impl_of<_Sender>, __copy_cvref_fn<_Sender>, __sexpr_uncurry_fn<_Fn>>;
-
-    template <class _Sender>
-    using __desc_of = __sexpr_uncurry<_Sender, __q<__desc>>;
-
-    using __get_desc = __sexpr_uncurry_fn<__q<__desc>>;
   } // namespace __detail
 
   namespace {
@@ -93,34 +69,6 @@ namespace stdexec {
 #  define STDEXEC_SEXPR_DESCRIPTOR(_Tag, _Data, _Child)                                            \
     stdexec::__descriptor_fn_v<stdexec::__detail::__desc<_Tag, _Data, _Child>>
 #endif
-
-  template <class _Sender>
-  using tag_of_t = typename __detail::__desc_of<_Sender>::__tag;
-
-  template <class _Sender>
-  using __data_of = typename __detail::__desc_of<_Sender>::__data;
-
-  template <class _Sender, class _Continuation = __q<__types>>
-  using __children_of = //
-    __mapply<_Continuation, typename __detail::__desc_of<_Sender>::__children>;
-
-  template <class _Ny, class _Sender>
-  using __nth_child_of = __children_of<_Sender, __mbind_front_q<__m_at, _Ny>>;
-
-  template <std::size_t _Ny, class _Sender>
-  using __nth_child_of_c = __children_of<_Sender, __mbind_front_q<__m_at, __msize_t<_Ny>>>;
-
-  template <class _Sender>
-  using __child_of = __children_of<_Sender, __q<__mfront>>;
-
-  template <class _Sender>
-  inline constexpr std::size_t __nbr_children_of = __v<__children_of<_Sender, __msize>>;
-
-  template <class _Fn, class _Tp>
-    requires __mvalid<tag_of_t, _Tp> && __mvalid<__detail::__sexpr_uncurry, _Tp, _Fn>
-  struct __uncurry_<_Fn, _Tp> {
-    using __t = __detail::__sexpr_uncurry<_Tp, _Fn>;
-  };
 
   template <class _Tag>
   struct __sexpr_impl;
@@ -456,11 +404,6 @@ namespace stdexec {
       };
     }
 
-    template <class _Sender>
-    concept __non_dependent_sender = //
-      requires { typename _Sender::completion_signatures; }
-      || requires { requires _Sender::__is_non_dependent(); };
-
     template <class _Tag, class... _Child>
     concept __is_non_dependent_sexpr = //
       !requires { typename __sexpr_impl<_Tag>::is_dependent; }
@@ -611,34 +554,6 @@ namespace stdexec {
 
   template <class _Tag>
   inline constexpr __detail::__make_sexpr_t<_Tag> __make_sexpr{};
-
-  namespace __detail {
-    struct __sexpr_apply_t {
-      template <class _Sender, class _ApplyFn>
-      STDEXEC_ATTRIBUTE((always_inline))
-      auto
-        operator()(_Sender&& __sndr, _ApplyFn&& __fun) const //
-        noexcept(
-          noexcept(__sndr.apply(static_cast<_Sender&&>(__sndr), static_cast<_ApplyFn&&>(__fun)))) //
-        -> decltype(__sndr.apply(static_cast<_Sender&&>(__sndr), static_cast<_ApplyFn&&>(__fun))) {
-        return __sndr.apply(static_cast<_Sender&&>(__sndr), static_cast<_ApplyFn&&>(__fun)); //
-      }
-    };
-  } // namespace __detail
-
-  using __detail::__sexpr_apply_t;
-  inline constexpr __sexpr_apply_t __sexpr_apply{};
-
-  template <class _Sender, class _ApplyFn>
-  using __sexpr_apply_result_t = __call_result_t<__sexpr_apply_t, _Sender, _ApplyFn>;
-
-  template <class _Sender>
-  concept sender_expr = //
-    __mvalid<tag_of_t, _Sender>;
-
-  template <class _Sender, class _Tag>
-  concept sender_expr_for = //
-    sender_expr<_Sender> && same_as<tag_of_t<_Sender>, _Tag>;
 
   // The __name_of utility defined below is used to pretty-print the type names of
   // senders in compiler diagnostics.
