+++ include/stdexec/__detail/__senders.hpp
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) 2021-2022 NVIDIA Corporation
+ *
+ * Licensed under the Apache License Version 2.0 with LLVM Exceptions
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *   https://llvm.org/LICENSE.txt
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#pragma once
+
+#include "__execution_fwd.hpp"
+
+// include these after __execution_fwd.hpp
+#include "__completion_signatures.hpp"
+#include "__concepts.hpp"
+#include "__env.hpp"
+#include "__receivers.hpp"
+#include "__type_traits.hpp"
+
+namespace stdexec {
+  /////////////////////////////////////////////////////////////////////////////
+  // [execution.senders]
+  struct sender_t {
+    using sender_concept = sender_t;
+  };
+
+  namespace __detail {
+    template <class _Sender>
+    concept __non_dependent_sender = //
+      requires { typename _Sender::completion_signatures; }
+      || requires { requires _Sender::__is_non_dependent(); };
+
+    template <class _Sender>
+    concept __enable_sender = //
+      derived_from<typename _Sender::sender_concept, sender_t>
+      || requires { typename _Sender::is_sender; } // NOT TO SPEC back compat
+      || __awaitable<_Sender, __env::__promise<empty_env>>;
+  } // namespace __detail
+
+  template <class _Sender>
+  inline constexpr bool enable_sender = __detail::__enable_sender<_Sender>;
+
+  template <class _Sender>
+  concept sender = enable_sender<__decay_t<_Sender>> &&                  //
+                   environment_provider<__cref_t<_Sender>> &&            //
+                   __detail::__consistent_completion_domains<_Sender> && //
+                   move_constructible<__decay_t<_Sender>> &&             //
+                   constructible_from<__decay_t<_Sender>, _Sender>;
+
+  template <class _Sender, class... _Env>
+  concept sender_in =         //
+    (sizeof...(_Env) <= 1) && //
+    sender<_Sender> &&        //
+    requires(_Sender&& __sndr, _Env&&... __env) {
+      {
+        get_completion_signatures(static_cast<_Sender&&>(__sndr), static_cast<_Env&&>(__env)...)
+      } -> __valid_completion_signatures;
+    };
+
+  /////////////////////////////////////////////////////////////////////////////
+  // early sender type-checking
+  template <class _Sender>
+  concept __well_formed_sender = //
+    !__detail::__non_dependent_sender<_Sender>
+    || __valid_completion_signatures<__completion_signatures_of_t<_Sender>>;
+
+} // namespace stdexec
