+++ include/stdexec/__detail/__transform_sender.hpp
@@ -18,8 +18,9 @@
 #include "__execution_fwd.hpp"
 
 // include these after __execution_fwd.hpp
-#include "__basic_sender.hpp"
+#include "__sender_introspection.hpp"
 #include "__concepts.hpp"
+#include "__diagnostics.hpp"
 #include "__domain.hpp"
 #include "__type_traits.hpp"
